pyproject.toml
[project]
name = "agentleh-app"
version = "0.1.0"
description = "Agentleh web app — self-serve agent onboarding"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "psycopg[binary]>=3.2",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.17",
    "dynaconf>=3.2",
    "httpx>=0.28",
]

[project.scripts]
dev = "scripts.dev:dev"
backend = "scripts.dev:backend"
test = "scripts.dev:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "lib", "services", "scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
